feature: split out meter report creator functionality
