+++ v2/controllers/marketplace/meterdefinition_controller_test.go
@@ -69,7 +69,7 @@ var _ = Describe("MeterDefinitionController", func() {
 			Log:     log,
 			cc:      reconcileutils.NewClientCommand(r.Client, s, log),
 			patcher: patch.RHMDefaultPatcher,
-			cfg:     cfg2,
+			cfg:     operatorCfg,
 		}
 		return nil
 	}
